Add files via upload
Ex11.1. Uses a recursive subroutine to print out list of files in a directory, and list of files in a file which can be a directory in the directory. Typing the recursive subroutine was quite refreshing.

Ex 11.2. Counts the number of lines in each file specified on the command line arguments. Uses Scanner to process each file. Take a look.

Ex 11.3. A Network server program that delivers the list of files in the server after the connection has been successful. Knew that I had learnt how to create a connection between two computers. Runs from cmd.

Ex 11.4. Runs from cmd too. Receives the files sent from the previous exercise and displays them to the user using specified commands. Meant to be as a simple GUI, but I had designed the backbone as the command prompt to test the previous exercise. Enjoy reading.

Ex 11.5. This exercise tests usage on XML. Stores a phone directory in a file using xml, and when the program is loaded again, it can display the names stored parsing the xml file. Coded it once without errors to my surprise.